move IMDiscountSchemaBL.calcualtePrice to testing
https://github.com/metasfresh/metasfresh/issues/3891
