+++ de.metas.business/src/main/java/org/adempiere/pricing/api/IMDiscountSchemaBL.java
@@ -44,13 +44,6 @@ public interface IMDiscountSchemaBL extends ISingletonService
 	DiscountResult calculateDiscount(final CalculateDiscountRequest request);
 
 	/**
-	 * Calculate Discounted Price
-	 *
-	 * @return
-	 */
-	BigDecimal calculatePrice(final CalculateDiscountRequest request);
-
-	/**
 	 * Criteria apply
 	 *
 	 * @param br
